refactor b2Math for performance & maintainability
