+++ Box2D/Box2D/Collision/b2TimeOfImpact.ts
@@ -17,7 +17,7 @@
 */
 
 import { b2_linearSlop, b2_maxPolygonVertices } from "../Common/b2Settings";
-import { b2Abs, b2Max, b2Vec2, b2Rot, b2Transform, b2Sweep } from "../Common/b2Math";
+import { b2Vec2, b2Rot, b2Transform, b2Sweep } from "../Common/b2Math";
 import { b2Timer } from "../Common/b2Timer";
 import { b2Distance, b2DistanceInput, b2DistanceOutput, b2DistanceProxy, b2SimplexCache } from "./b2Distance";
 
@@ -288,7 +288,7 @@ export function b2TimeOfImpact(output: b2TOIOutput, input: b2TOIInput): void {
   const tMax: number = input.tMax;
 
   const totalRadius: number = proxyA.m_radius + proxyB.m_radius;
-  const target: number = b2Max(b2_linearSlop, totalRadius - 3 * b2_linearSlop);
+  const target: number = Math.max(b2_linearSlop, totalRadius - 3 * b2_linearSlop);
   const tolerance: number = 0.25 * b2_linearSlop;
   ///b2Assert(target > tolerance);
 
@@ -431,7 +431,7 @@ export function b2TimeOfImpact(output: b2TOIOutput, input: b2TOIInput): void {
 
         const s: number = fcn.Evaluate(indexA[0], indexB[0], t);
 
-        if (b2Abs(s - target) < tolerance) {
+        if (Math.abs(s - target) < tolerance) {
           // t2 holds a tentative value for t1
           t2 = t;
           break;
@@ -451,7 +451,7 @@ export function b2TimeOfImpact(output: b2TOIOutput, input: b2TOIInput): void {
         }
       }
 
-      b2_toiMaxRootIters = b2Max(b2_toiMaxRootIters, rootIterCount);
+      b2_toiMaxRootIters = Math.max(b2_toiMaxRootIters, rootIterCount);
 
       ++pushBackIter;
 
@@ -475,9 +475,9 @@ export function b2TimeOfImpact(output: b2TOIOutput, input: b2TOIInput): void {
     }
   }
 
-  b2_toiMaxIters = b2Max(b2_toiMaxIters, iter);
+  b2_toiMaxIters = Math.max(b2_toiMaxIters, iter);
 
   const time: number = timer.GetMilliseconds();
-  b2_toiMaxTime = b2Max(b2_toiMaxTime, time);
+  b2_toiMaxTime = Math.max(b2_toiMaxTime, time);
   b2_toiTime += time;
 }
