+++ Box2D/Box2D/Dynamics/Joints/b2WheelJoint.ts
@@ -17,7 +17,7 @@
 */
 
 import { b2_pi, b2_linearSlop } from "../../Common/b2Settings";
-import { b2Abs, b2Clamp, b2Vec2, b2Rot } from "../../Common/b2Math";
+import { b2Clamp, b2Vec2, b2Rot } from "../../Common/b2Math";
 import { b2Joint, b2JointDef, b2JointType } from "./b2Joint";
 import { b2SolverData } from "../b2TimeStep";
 import { b2Body } from "../b2Body";
@@ -427,7 +427,7 @@ export class b2WheelJoint extends b2Joint {
     // data.positions[this.m_indexB].c = cB;
     data.positions[this.m_indexB].a = aB;
 
-    return b2Abs(C) <= b2_linearSlop;
+    return Math.abs(C) <= b2_linearSlop;
   }
 
   public GetDefinition(def: b2WheelJointDef): b2WheelJointDef {
