+++ Box2D/Box2D/Particle/b2ParticleSystem.ts
@@ -20,7 +20,7 @@
 
 import { b2_linearSlop, b2_maxFloat, b2_maxParticleIndex, b2_invalidParticleIndex, b2_minParticleSystemBufferCapacity, b2_maxTriadDistanceSquared, b2_barrierCollisionTime, b2MakeArray } from "../Common/b2Settings";
 import { b2_maxParticlePressure, b2_minParticleWeight, b2_maxParticleForce, b2_particleStride } from "../Common/b2Settings";
-import { b2Min, b2Max, b2Abs, b2Clamp, b2Sqrt, b2InvSqrt, b2Vec2, b2Rot, b2Transform } from "../Common/b2Math";
+import { b2Clamp, b2InvSqrt, b2Vec2, b2Rot, b2Transform } from "../Common/b2Math";
 import { b2Color } from "../Common/b2Draw";
 import { b2AABB, b2RayCastInput, b2RayCastOutput } from "../Collision/b2Collision";
 import { b2ShapeType, b2Shape, b2MassData } from "../Collision/Shapes/b2Shape";
@@ -328,7 +328,7 @@ export class b2ParticleContact {
   ApproximatelyEqual(rhs: b2ParticleContact): boolean {
     const MAX_WEIGHT_DIFF = 0.01; // Weight 0 ~ 1, so about 1%
     const MAX_NORMAL_DIFF_SQ = 0.01 * 0.01; // Normal length = 1, so 1%
-    return this.indexA === rhs.indexA && this.indexB === rhs.indexB && this.flags === rhs.flags && b2Abs(this.weight - rhs.weight) < MAX_WEIGHT_DIFF && b2Vec2.DistanceSquaredVV(this.normal, rhs.normal) < MAX_NORMAL_DIFF_SQ;
+    return this.indexA === rhs.indexA && this.indexB === rhs.indexB && this.flags === rhs.flags && Math.abs(this.weight - rhs.weight) < MAX_WEIGHT_DIFF && b2Vec2.DistanceSquaredVV(this.normal, rhs.normal) < MAX_NORMAL_DIFF_SQ;
   }
 }
 
@@ -1854,7 +1854,7 @@ export class b2ParticleSystem {
       let p2 = b2Vec2.DotVV(p, p);
       let determinant = pv * pv - v2 * (p2 - this.m_squaredDiameter);
       if (determinant >= 0) {
-        let sqrtDeterminant = b2Sqrt(determinant);
+        let sqrtDeterminant = Math.sqrt(determinant);
         // find a solution between 0 and fraction
         let t = (-pv - sqrtDeterminant) / v2;
         if (t > fraction) {
@@ -1871,7 +1871,7 @@ export class b2ParticleSystem {
         n.Normalize();
         ///float32 f = callback.ReportParticle(this, i, point1 + t * v, n, t);
         let f = callback.ReportParticle(this, i, b2Vec2.AddVMulSV(point1, t, v, s_point), n, t);
-        fraction = b2Min(fraction, f);
+        fraction = Math.min(fraction, f);
         if (fraction <= 0) {
           break;
         }
@@ -2288,7 +2288,7 @@ export class b2ParticleSystem {
           pair.indexA = a;
           pair.indexB = b;
           pair.flags = contact.flags;
-          pair.strength = b2Min(
+          pair.strength = Math.min(
             groupA ? groupA.m_strength : 1,
             groupB ? groupB.m_strength : 1);
           ///pair.distance = b2Distance(pos_data[a], pos_data[b]); // TODO: this was wrong!
@@ -2350,7 +2350,7 @@ export class b2ParticleSystem {
           triad.indexB = b;
           triad.indexC = c;
           triad.flags = af | bf | cf;
-          triad.strength = b2Min(b2Min(
+          triad.strength = Math.min(Math.min(
               groupA ? groupA.m_strength : 1,
               groupB ? groupB.m_strength : 1),
             groupC ? groupC.m_strength : 1);
@@ -2633,8 +2633,8 @@ export class b2ParticleSystem {
     // The number of iterations is equal to particle number from the deepest
     // particle to the nearest surface particle, and in general it is smaller
     // than sqrt of total particle number.
-    ///int32 iterationCount = (int32)b2Sqrt((float)m_count);
-    let iterationCount = b2Sqrt(this.m_count) >> 0;
+    ///int32 iterationCount = (int32)Math.sqrt((float)m_count);
+    let iterationCount = Math.sqrt(this.m_count) >> 0;
     for (let t = 0; t < iterationCount; t++) {
       let updated = false;
       for (let k = 0; k < contactGroupsCount; k++) {
@@ -3099,12 +3099,12 @@ export class b2ParticleSystem {
       ///let p2 = p1 + step.dt * v;
       let p2_x = p1.x + step.dt * v.x;
       let p2_y = p1.y + step.dt * v.y;
-      ///aabb.lowerBound = b2Min(aabb.lowerBound, b2Min(p1, p2));
-      aabb.lowerBound.x = b2Min(aabb.lowerBound.x, b2Min(p1.x, p2_x));
-      aabb.lowerBound.y = b2Min(aabb.lowerBound.y, b2Min(p1.y, p2_y));
-      ///aabb.upperBound = b2Max(aabb.upperBound, b2Max(p1, p2));
-      aabb.upperBound.x = b2Max(aabb.upperBound.x, b2Max(p1.x, p2_x));
-      aabb.upperBound.y = b2Max(aabb.upperBound.y, b2Max(p1.y, p2_y));
+      ///aabb.lowerBound = Math.min(aabb.lowerBound, Math.min(p1, p2));
+      aabb.lowerBound.x = Math.min(aabb.lowerBound.x, Math.min(p1.x, p2_x));
+      aabb.lowerBound.y = Math.min(aabb.lowerBound.y, Math.min(p1.y, p2_y));
+      ///aabb.upperBound = Math.max(aabb.upperBound, Math.max(p1, p2));
+      aabb.upperBound.x = Math.max(aabb.upperBound.x, Math.max(p1.x, p2_x));
+      aabb.upperBound.y = Math.max(aabb.upperBound.y, Math.max(p1.y, p2_y));
     }
     let callback = new b2ParticleSystem.SolveCollisionCallback(this, step);
     this.m_world.QueryAABB(callback, aabb);
@@ -3118,8 +3118,8 @@ export class b2ParticleSystem {
       let v = vel_data[i];
       let v2 = b2Vec2.DotVV(v, v);
       if (v2 > criticalVelocitySquared) {
-        ///v *= b2Sqrt(criticalVelocitySquared / v2);
-        v.SelfMul(b2Sqrt(criticalVelocitySquared / v2));
+        ///v *= Math.sqrt(criticalVelocitySquared / v2);
+        v.SelfMul(Math.sqrt(criticalVelocitySquared / v2));
       }
     }
   }
@@ -3170,9 +3170,9 @@ export class b2ParticleSystem {
         let pb = pos_data[b];
         /// b2AABB aabb;
         let aabb = s_aabb;
-        ///aabb.lowerBound = b2Min(pa, pb);
+        ///aabb.lowerBound = Math.min(pa, pb);
         b2Vec2.MinV(pa, pb, aabb.lowerBound);
-        ///aabb.upperBound = b2Max(pa, pb);
+        ///aabb.upperBound = Math.max(pa, pb);
         b2Vec2.MaxV(pa, pb, aabb.upperBound);
         let aGroup = this.m_groupBuffer[a];
         let bGroup = this.m_groupBuffer[b];
@@ -3222,7 +3222,7 @@ export class b2ParticleSystem {
             } else {
               let det = e1 * e1 - 4 * e0 * e2;
               if (det < 0) continue;
-              let sqrtDet = b2Sqrt(det);
+              let sqrtDet = Math.sqrt(det);
               let t1 = (-e1 - sqrtDet) / (2 * e2);
               let t2 = (-e1 + sqrtDet) / (2 * e2);
               ///if (t1 > t2) b2Swap(t1, t2);
@@ -3378,8 +3378,8 @@ export class b2ParticleSystem {
     let maxPressure = b2_maxParticlePressure * criticalPressure;
     for (let i = 0; i < this.m_count; i++) {
       let w = this.m_weightBuffer[i];
-      let h = pressurePerWeight * b2Max(0.0, w - b2_minParticleWeight);
-      this.m_accumulationBuffer[i] = b2Min(h, maxPressure);
+      let h = pressurePerWeight * Math.max(0.0, w - b2_minParticleWeight);
+      this.m_accumulationBuffer[i] = Math.min(h, maxPressure);
     }
     // ignores particles which have their own repulsive force
     if (this.m_allParticleFlags & b2ParticleSystem.k_noPressureFlags) {
@@ -3454,7 +3454,7 @@ export class b2ParticleSystem {
       let v = b2Vec2.SubVV(b.GetLinearVelocityFromWorldPoint(p, b2Vec2.s_t0), vel_data[a], s_v);
       let vn = b2Vec2.DotVV(v, n);
       if (vn < 0) {
-        let damping = b2Max(linearDamping * w, b2Min(-quadraticDamping * vn, 0.5));
+        let damping = Math.max(linearDamping * w, Math.min(-quadraticDamping * vn, 0.5));
         ///b2Vec2 f = damping * m * vn * n;
         let f = b2Vec2.MulSV(damping * m * vn, n, s_f);
         ///m_velocityBuffer.data[a] += GetParticleInvMass() * f;
@@ -3473,8 +3473,8 @@ export class b2ParticleSystem {
       let v = b2Vec2.SubVV(vel_data[b], vel_data[a], s_v);
       let vn = b2Vec2.DotVV(v, n);
       if (vn < 0) {
-        ///float32 damping = b2Max(linearDamping * w, b2Min(- quadraticDamping * vn, 0.5f));
-        let damping = b2Max(linearDamping * w, b2Min(-quadraticDamping * vn, 0.5));
+        ///float32 damping = Math.max(linearDamping * w, Math.min(- quadraticDamping * vn, 0.5f));
+        let damping = Math.max(linearDamping * w, Math.min(-quadraticDamping * vn, 0.5));
         ///b2Vec2 f = damping * vn * n;
         let f = b2Vec2.MulSV(damping * vn, n, s_f);
         ///this.m_velocityBuffer.data[a] += f;
@@ -3522,8 +3522,8 @@ export class b2ParticleSystem {
           // Calculate b.m_I from public functions of b2Body.
           ///this.InitDampingParameter(&invMassB, &invInertiaB, &tangentDistanceB, b.GetMass(), b.GetInertia() - b.GetMass() * b.GetLocalCenter().LengthSquared(), b.GetWorldCenter(), p, n);
           this.InitDampingParameter(invMassB, invInertiaB, tangentDistanceB, b.GetMass(), b.GetInertia() - b.GetMass() * b.GetLocalCenter().LengthSquared(), b.GetWorldCenter(), p, n);
-          ///float32 f = damping * b2Min(w, 1.0) * this.ComputeDampingImpulse(invMassA, invInertiaA, tangentDistanceA, invMassB, invInertiaB, tangentDistanceB, vn);
-          let f = damping * b2Min(w, 1.0) * this.ComputeDampingImpulse(invMassA[0], invInertiaA[0], tangentDistanceA[0], invMassB[0], invInertiaB[0], tangentDistanceB[0], vn);
+          ///float32 f = damping * Math.min(w, 1.0) * this.ComputeDampingImpulse(invMassA, invInertiaA, tangentDistanceA, invMassB, invInertiaB, tangentDistanceB, vn);
+          let f = damping * Math.min(w, 1.0) * this.ComputeDampingImpulse(invMassA[0], invInertiaA[0], tangentDistanceA[0], invMassB[0], invInertiaB[0], tangentDistanceB[0], vn);
           ///this.ApplyDamping(invMassA, invInertiaA, tangentDistanceA, true, aGroup, a, f, n);
           this.ApplyDamping(invMassA[0], invInertiaA[0], tangentDistanceA[0], true, aGroup, a, f, n);
           ///b.ApplyLinearImpulse(-f * n, p, true);
@@ -3823,7 +3823,7 @@ export class b2ParticleSystem {
         let h = this.m_weightBuffer[a] + this.m_weightBuffer[b];
         ///b2Vec2 s = m_accumulation2Buffer[b] - m_accumulation2Buffer[a];
         let s = b2Vec2.SubVV(this.m_accumulation2Buffer[b], this.m_accumulation2Buffer[a], s_s);
-        let fn = b2Min(
+        let fn = Math.min(
           pressureStrength * (h - 2) + normalStrength * b2Vec2.DotVV(s, n),
           maxVelocityVariation) * w;
         ///b2Vec2 f = fn * n;
@@ -4156,8 +4156,8 @@ export class b2ParticleSystem {
       for (let i = group.m_firstIndex; i < group.m_lastIndex; i++) {
         let j = newIndices[i];
         if (j >= 0) {
-          firstIndex = b2Min(firstIndex, j);
-          lastIndex = b2Max(lastIndex, j + 1);
+          firstIndex = Math.min(firstIndex, j);
+          lastIndex = Math.max(lastIndex, j + 1);
         } else {
           modified = true;
         }
